project/requirements.txt
alembic==1.11.1
bcrypt==4.0.1
email-validator==2.0.0.post2
fastapi==0.100.0
httpx==0.24.1
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==2.0.3
pydantic-settings==2.0.1
python-multipart==0.0.6
pytest==7.4.0
python-jose==3.3.0
sqlalchemy==2.0.18
uvicorn==0.22.0
